Move clearWorksheet to /lib
