+++ lib/clearWorksheet.js
@@ -1,4 +1,4 @@
-import { getGraphClient } from "./lib/msAuth.js"; // Import the authentication functions
+import { getGraphClient } from "./msAuth.js"; // Import the authentication functions
 
 // Function to clear the contents of a specified range in a worksheet
 export const clearWorksheetRange = async (
