+++ lib/worksheetProcessing.js
@@ -7,7 +7,7 @@ import {
   getCellRange,
 } from "./utils.js";
 import { getFileIdByName } from "./oneDrive.js";
-import { clearWorksheetRange } from "../clearWorksheet.js";
+import { clearWorksheetRange } from "./clearWorksheet.js";
 import { copyFileInOneDrive } from "../copyFile.js";
 // import { copyFileInOneDrive } from "../copyFile.js";
 // import { clearWorksheetRange } from "../clearWorksheet.js";
